% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLegend}
\alias{plotLegend}
\title{Plot the legend}
\usage{
plotLegend(appearances)
}
\arguments{
\item{appearances}{list of data frame(s) specifying the appearances of each type of temporalities; each data frame should have a column named "color" specifying the color used for each type of urban temporality; moreover, if there is no column named "leg", then row names will be used as labels}
}
\description{
Plot the legend associated with urban temporalities, in its on plot.
}
\seealso{
\code{\link{plotTemporalities}}
}
\author{
Timothee Flutre
}
