% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTemporalities}
\alias{plotTemporalities}
\title{Plot urban temporalities}
\usage{
plotTemporalities(temporalities, appearances, main = "Urban temporalities",
  data.source = "author, date", vertical.bars.per.hour = FALSE,
  verbose = 1)
}
\arguments{
\item{temporalities}{list of data frame(s) containing urban temporalities, as returned by \code{\link[=readTemporalities]{readTemporalities()}}}

\item{appearances}{list of data.frame(s) specifying the appearances of each type of temporalities}

\item{main}{main title}

\item{data.source}{source of the data, for instance \code{"M. Gibert, 2013"} (skipped if NULL)}

\item{vertical.bars.per.hour}{if TRUE, a vertical bar is added for each hour}

\item{verbose}{verbosity level (0/1)}
}
\description{
Draws a plot of urban temporalities, similarly to figure 4.6 in Marie Gibert's \href{https://www.academia.edu/7549254}{PhD thesis} (2014).
}
\author{
Timothee Flutre
}
