% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readAppearances}
\alias{readAppearances}
\title{Input of appearances}
\usage{
readAppearances(file, verbose = 1)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
The content of this file should correspond to urban temporalities as read by \code{\link[=readTemporalities]{readTemporalities()}}, with one row per event.
The first line (header) should contain column names, and columns should be separated by a tabulation:
\itemize{
\item first column: should be named \code{"space"};
\item second column: should be named \code{"event"};
\item third column: should be named \code{"color"};
\item (optional) fourth column: should be named \code{"density"}.
}}

\item{verbose}{verbosity level (0/1/2)}
}
\value{
list of data frame(s), one per space type
}
\description{
Reads a file containing the appearances in table format, and returns a list of data frame(s) from it.
}
\seealso{
\code{\link[=readTemporalities]{readTemporalities()}}, \code{\link[=plotTemporalities]{plotTemporalities()}}
}
\author{
Timothee Flutre
}
