% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readTemporalities}
\alias{readTemporalities}
\title{Input of urban temporalities}
\usage{
readTemporalities(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
The first line (header) should contain column names.
Columns should be separated by a tabulation.}
}
\value{
data frame
}
\description{
Reads a file containing urban temporalities in table format and creates a data frame from it.
}
\author{
Timothee Flutre
}
